requirements.txt
asgiref==3.8.1
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.4
defusedxml==0.7.1
Django==5.2.3
django-bootstrap5==25.1
django-filter==25.1
django-taggit==6.1.0
djangorestframework==3.16.0
idna==3.10
Markdown==3.8
oauthlib==3.2.2
pillow==11.2.1
psycopg==3.2.9
psycopg-binary==3.2.9
pycparser==2.22
PyJWT==2.10.1
python-dotenv==1.0.0
python3-openid==3.2.0
requests==2.32.4
requests-oauthlib==2.0.0
social-auth-app-django==5.4.3
social-auth-core==4.6.1
sqlparse==0.5.3
typing_extensions==4.14.0
urllib3==2.4.0
